feat: add exe time to process result
This adds the time it took to run a process to its result, and
switches from an int to a struct to support this.

This lets us measure the time it took to run commands from a test
runner without including the round trip time from the test runner to
the node.